Break apart MenuItem to partial classes
#50
