+++ ConsoleMenu/Menu/MenuItem.cs
@@ -1,23 +1,14 @@
 ﻿using ConsoleMenu.Components;
-using ConsoleMenu.Events;
-using ConsoleMenu.Extensions;
 using ConsoleMenu.Managers;
-using ConsoleMenu.PerformAction;
 using ConsoleMenu.Render;
 using System.Diagnostics;
 
 namespace ConsoleMenu;
-public class MenuItem : IMenuItem
+public partial class MenuItem : IMenuItem
 {
-    private readonly ChildrenManager _childrenManager;
-    private bool _isCurrentlyVisible;
-    private List<IComponent> _components;
-
-    public Vector2 Position { get; set; }
     public IMenuItem Parent { get; set; }
-    public bool IsVisible { get; set; }
-    public bool MayCollapse { get; set; }
-    public event Action<IMenuItem, ConsoleKeyInfo> OnKeyPressed;
+    public Vector2 Position { get; set; }
+    public IContent Content { get; private set; }
 
     public MenuItem(string title)
     {
@@ -33,187 +24,4 @@ public class MenuItem : IMenuItem
         _components = new List<IComponent>();
         Debug.WriteLine($"MenuItem: '{Content.Title}' have now been created.", "Ctor");
     }
-
-    public Vector2 AreaNeeded()
-    {
-        if ((_isCurrentlyVisible == false || IsVisible == false) && MayCollapse)
-        {
-            return Vector2.ZERO;
-        }
-        Vector2 contentArea = Content.AreaNeeded();
-        Vector2 childrenArea = _childrenManager.AreaNeeded();
-        return contentArea.MaxAdd_Vertical(childrenArea);
-    }
-
-    public void SetRenderer<T>() where T : IContentRenderer, new()
-    {
-        IContentRenderer contentRenderer = new T();
-        if (Content is Content c && c.IsCurrentlyVisible)
-        {
-            Content.EraseContent();
-            Content.SetRenderer(contentRenderer.Render, contentRenderer.AreaNeeded);
-            Content.Render();
-            return;
-        }
-        Content.SetRenderer(contentRenderer.Render, contentRenderer.AreaNeeded);
-        Debug.WriteLine($"MenuItem: '{Content.Title}' have now changed renderer.", "SetRenderer");
-    }
-
-    public void Render()
-    {
-        if (IsVisible == false)
-        {
-            return;
-        }
-        _isCurrentlyVisible = true;
-
-        Content.Render();
-        Vector2 areaNeeded = Content.AreaNeeded();
-        _childrenManager.PositionOfFirstChild = new Vector2(Position.X, Position.Y + areaNeeded.Y);
-        _childrenManager.Render();
-        Debug.WriteLine($"MenuItem: '{Content.Title}' have now been Rendered", "Render");
-    }
-
-    public void EraseContent()
-    {
-        if (_isCurrentlyVisible)
-        {
-            Content.EraseContent();
-            _childrenManager.EraseContent();
-            _isCurrentlyVisible = false;
-        }
-        Debug.WriteLine($"MenuItem: '{Content.Title}' have now been erased.", "EraseContent");
-    }
-
-    internal MenuItem GetRoot(MenuItem menuItem)
-    {
-        if (menuItem.Parent is null)
-        {
-            return menuItem;
-        }
-        else if (menuItem.Parent is MenuItem parent)
-        {
-            return menuItem.GetRoot(parent);
-        }
-        else
-        {
-            throw new ArgumentException();
-        }
-    }
-
-    public void ReRender()
-    {
-        IMenuItem root = GetRoot(this);
-        root.EraseContent();
-        root.Render();
-    }
-
-    public IContent Content { get; private set; }
-
-    #region Components
-    public void AddComponent(IComponent component)
-    {
-        component.Owner = this;
-        _components.Add(component);
-    }
-    public IEnumerable<TComponent> GetComponents<TComponent>() where TComponent : IComponent
-    {
-        foreach (TComponent component in _components.OfType<TComponent>())
-        {
-            yield return component;
-        }
-    }
-    public IEnumerable<T> Values<T>()
-    {
-        IEnumerable<T> values = GetComponents<ValueComponent<T>>().Select(x => x.Value);
-        foreach (T component in values)
-        {
-            yield return component;
-        }
-    }
-
-    public void RemoveComponent(IComponent component)
-    {
-        _components.Remove(component);
-    }
-    #endregion
-
-    #region Children
-
-    public Orientation OrientationOfChildren
-    {
-        get => _childrenManager.OrientationOfChildren;
-        set => _childrenManager.OrientationOfChildren = value;
-    }
-    public bool IsChildrenVisible { get => _childrenManager.IsVisible; set => _childrenManager.IsVisible = value; }
-    public IMenuItem this[int i] => _childrenManager.GetChild(i);
-    public IMenuItem this[string s] => _childrenManager.GetChildren().FirstOrDefault(x => x.Content.Title == s);
-    public void AddChild(string title, int positionInList = int.MaxValue)
-    {
-        IMenuItem menuItem = new MenuItem(title);
-        AddChild(menuItem, positionInList);
-    }
-
-    public void AddChild<T>(T value, string title, int positionInList = int.MaxValue)
-    {
-        IMenuItem menuItem = new MenuItem(title);
-        menuItem.AddComponent(new ValueComponent<T>(value));
-        AddChild(menuItem, positionInList);
-    }
-    private void AddChild(IMenuItem menuItem, int positionInList)
-    {
-        if (positionInList != int.MaxValue)
-        {
-            menuItem.AddComponent(new ListPriorityComponent(positionInList));
-        }
-        _childrenManager.Add(menuItem);
-    }
-
-    public void RemoveChild(int i) => _childrenManager.Remove(i);
-    public void RemoveChild(IMenuItem menuItem) => _childrenManager.Remove(menuItem);
-    public IEnumerable<IMenuItem> GetChildren() => _childrenManager.GetChildren();
-    public bool HaveChildren() => _childrenManager.GetChildren().Any();
-    public Vector2 PositionOffsetOfFirstChild
-    {
-        get => _childrenManager.PositionOffsetOfFirstChild;
-        set => _childrenManager.PositionOffsetOfFirstChild = value;
-    }
-    public int PositionOffsetToNextChild
-    {
-        get => _childrenManager.PositionOffsetToNextChild;
-        set => _childrenManager.PositionOffsetToNextChild = value;
-    }
-    #endregion
-
-    #region Selection
-    public event Action<SelectionChangedEvent> OnSelectionChanged
-    {
-        add => _childrenManager.Selection.OnSelectionChanged += value;
-        remove => _childrenManager.Selection.OnSelectionChanged -= value;
-    }
-    public event Action<SelectionRenderedEvent> OnSelectionRendered
-    {
-        add => _childrenManager.Selection.OnSelectionRendered += value;
-        remove => _childrenManager.Selection.OnSelectionRendered -= value;
-    }
-    public IMenuItem GetSelectedChild() => _childrenManager.Selection.GetSelectedChild();
-    public bool IncrementSelection() => _childrenManager.Selection.Increment();
-    public bool DecrementSelection() => _childrenManager.Selection.Decrement();
-    #endregion
-
-
-    public bool KeyPressed(ConsoleKeyInfo key)
-    {
-        Debug.WriteLine($"MenuItem: '{Content.Title}' is about to process pressed keys", "KeyPressed");
-        if (HaveChildren())
-        {
-            if (GetSelectedChild().KeyPressed(key))
-            {
-                return true;
-            }
-        }
-        return ActionToPerform.MoveSelection(key, this);
-    }
-
-    public void PerformAction(ConsoleKeyInfo key) => OnKeyPressed?.Invoke(this, key);
 }
